# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Genisys2 Kanban API
  description: |
    Slice 4 "Live Board" Kanban API.

    Provides board snapshot retrieval, card CRUD, card moves between columns,
    processor callback handling, real-time SSE board streams, and an immutable
    audit log query endpoint. All successful JSON responses are wrapped in a
    `{ data: { ... } }` envelope.

    ## Deterministic Demo Board
    The API seeds a deterministic demo board on startup via `seedBoard`. The demo board
    includes a **Processing** column (`in-review`) gated by the `manager-approval` processor.

    ## Processing Columns
    When a card enters a `Processing` column, the card's `processing_state` becomes
    `PROCESSING`, `is_editable` becomes `false`, and the system dispatches an async
    `on-enter` hook to the column's processor. The processor must POST back to the
    provided `callback_url` to unlock the card or move it to another column.

    ## Real-Time SSE Stream
    `GET /boards/{boardId}/stream` opens a Server-Sent Events connection. It emits
    `CARD_CREATED`, `CARD_UPDATED`, `CARD_MOVED`, and `BOARD_RELOAD` events. Clients
    may supply a `Last-Event-ID` header to resume from a disconnect. The server buffers
    the last 5 minutes of events per board. If the cursor is older than the buffer,
    a `BOARD_RELOAD` event is emitted so the client can fetch a fresh snapshot.

    ## Audit Log
    `GET /boards/{boardId}/audit-log` queries the immutable `event_log` table. Every
...
servers:
...
paths:
...
  /boards/{boardId}/cards/{cardId}:
    get:
      operationId: getCard
      summary: Get a single card
      description: Retrieves a card by its UUID within the specified board.
      tags:
        - Cards
      parameters:
        - $ref: "#/components/parameters/BoardIdParam"
        - $ref: "#/components/parameters/CardIdParam"
      responses:
        "200":
          description: Card retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCardResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      operationId: updateCard
      summary: Update card title or description
      description: |
        Partially updates a card. At least one of `title` or `description` must be
        provided. The request must include the current `version` for optimistic locking.
        If the server's version has advanced, a `409 Conflict` is returned with the
        authoritative card state so the client can refresh and retry.

        Cards in `PROCESSING` or `ERROR` state (`is_editable: false`) cannot be edited.
      tags:
        - Cards
      parameters:
        - $ref: "#/components/parameters/BoardIdParam"
        - $ref: "#/components/parameters/CardIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCardRequest"
      responses:
        "200":
          description: Card updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateCardResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Optimistic locking conflict — card was modified by another user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardConflictResponse"
  /boards/{boardId}/cards/{cardId}/move:
    post:
      operationId: moveCard
      summary: Move a card to another column
      description: |
        Changes a card's `current_status` to the target column uid.

        **Blocked moves:** A card in `PROCESSING` state cannot be moved directly;
        the API returns `409 MOVE_BLOCKED`. The processor must call the callback
        endpoint with `move_to_column` to transition the card.

        **can-exit hook:** Before the move, a synchronous `can-exit` hook is dispatched
        to the column's processor with a strict 3-second timeout. If the processor
        returns `allowed: false`, the move is rejected with `409 Conflict`.
        The target column must exist in the board schema. The card `version` is
        incremented automatically on a successful move.
      tags:
        - Cards
      parameters:
        - $ref: "#/components/parameters/BoardIdParam"
        - $ref: "#/components/parameters/CardIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveCardRequest"
      responses:
        "200":
          description: Card moved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoveCardResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            Move blocked by processor `can-exit` hook, card is in `PROCESSING` state,
            or optimistic locking conflict.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoveCardBlockedResponse"
        "500":
          description: Internal move failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    CardEntity:
      type: object
      required:
        - uid
        - board_uid
        - display_id
        - title
        - description
        - version
        - processing_state
        - is_editable
        - payload
        - current_status
        - created_at
        - updated_at
      properties:
        uid:
          type: string
          format: uuid
          description: Card unique identifier
        board_uid:
          type: string
          format: uuid
          description: Parent board UUID
        display_id:
          type: string
          pattern: "^[A-Z][A-Z0-9]{0,9}-[1-9]\\d*$"
          description: Human-readable display ID (e.g. `S0-1`)
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Card title
        description:
          type: string
          maxLength: 20000
          nullable: true
          description: Optional card description
        version:
          type: integer
          minimum: 1
          description: Optimistic concurrency version (auto-incremented)
        processing_state:
          type: string
          enum:
            - IDLE
            - PROCESSING
            - ERROR
          description: |
            Current processing state. `PROCESSING` means an async hook is in flight.
            `ERROR` means the processor reported failure. Both lock the card
            (`is_editable: false`).
        is_editable:
          type: boolean
          description: |
            Whether the card is currently editable. Automatically `false` when
            `processing_state` is `PROCESSING` or `ERROR`.
        payload:
          type: object
          additionalProperties: true
          description: Arbitrary JSON payload
        current_status:
          type: string
          minLength: 1
          maxLength: 64
          pattern: "^[A-Za-z0-9:_-]+$"
          description: Current column uid
        created_at:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 last update timestamp
...
    GetCardResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - card
          properties:
            card:
              $ref: "#/components/schemas/CardEntity"
    UpdateCardRequest:
      type: object
      required:
        - version
      properties:
        version:
          type: integer
          minimum: 1
          description: Current optimistic-lock version from the card entity
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: New card title
        description:
          type: string
          maxLength: 20000
          nullable: true
          description: New card description
    UpdateCardResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - card
          properties:
            card:
              $ref: "#/components/schemas/CardEntity"
    MoveCardRequest:
      type: object
      required:
        - to_column_uid
      properties:
        to_column_uid:
          type: string
          minLength: 1
          maxLength: 64
          pattern: "^[A-Za-z0-9:_-]+$"
          description: Destination column uid
    MoveCardResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - card
          properties:
            card:
              $ref: "#/components/schemas/CardEntity"
...
    CardConflictResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - details
          properties:
            code:
              type: string
              enum:
                - CONFLICT
            message:
              type: string
              description: Human-readable conflict explanation
            details:
              type: object
              required:
                - current_version
                - card
              properties:
                current_version:
                  type: integer
                  minimum: 1
                card:
                  $ref: "#/components/schemas/CardEntity"
    MoveCardBlockedResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - MOVE_BLOCKED
            message:
              type: string
              description: Processor-provided reason for blocking the move
            details:
              type: object
              properties:
                hook:
                  type: string
                  enum:
                    - can-exit
...
    ApiError:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              minLength: 1
              description: Machine-readable error code
            message:
              type: string
              minLength: 1
              description: Human-readable error message
            details:
              type: object
              additionalProperties: true
              description: Optional structured error details
  parameters:
    BoardIdParam:
      name: boardId
      in: path
      required: true
      description: Board UUID
      schema:
        type: string
        format: uuid
    CardIdParam:
      name: cardId
      in: path
      required: true
      description: Card UUID
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error:
              code: INVALID_PARAMS
              message: Invalid path parameters
              details:
                issues:
                  - path: ["boardId"]
                    message: Invalid uuid
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error:
              code: BOARD_NOT_FOUND
              message: Board not found
...